refactor(ranking): extract reusable leaderboard results
Move podium and table rendering into a shared results component while keeping local profile editing opt-in for read-only consumers.
