+++ web/src/features/ranking/RankingPage.tsx
@@ -8,13 +8,12 @@ import { Modal } from '@/components/ui/Modal';
 import { QuestionMarkHelp } from '@/components/ui/QuestionMarkHelp';
 import { RankingApiError } from './api';
 import { RankingAvatar } from './components/RankingAvatar';
+import { RankingLeaderboardResults } from './components/RankingLeaderboardResults';
 import { RankingMetricSelect, RankingToolbar } from './components/RankingToolbar';
-import { formatLeaderboardValue, formatOverallMetricValue } from './format';
 import { normalizeRankingDisplayName, RANKING_DISPLAY_NAME_MAX_LENGTH, type RankingProfileError } from './profile';
 import type {
   LocalRankingProfileRequest,
   LocalRankingProfileResponse,
-  RankingDetailMetric,
   RankingLeaderboardEntry,
   RankingLeaderboardResponse,
   RankingMetadataResponse,
@@ -27,15 +26,6 @@ import type {
 import styles from './RankingPage.module.scss';
 
 const AVATAR_IDS = Array.from({ length: 66 }, (_, index) => index + 1);
-const OVERALL_METRICS: RankingDetailMetric[] = [
-  'total_tokens',
-  'request_count',
-  'cache_read_rate',
-  'ttft_average',
-  'latency_average',
-  'peak_tpm',
-  'peak_rpm',
-];
 
 type RankingAction = 'join' | 'sync' | 'pause' | 'resume' | 'exit' | null;
 type ProfileAction = Exclude<RankingAction, 'join' | null>;
@@ -701,8 +691,6 @@ function LeaderboardCard({
   language,
 }: LeaderboardCardProps) {
   const rows = useMemo(() => board?.entries.slice(0, 100) ?? [], [board]);
-  const podium = rows.slice(0, 3);
-  const tableRows = rows;
   const scoreExplanation = resolveScoreExplanation(scoreExplanationBoard, metric, language);
   const hasRankingProfile = status?.status === 'active' || status?.status === 'paused';
   const profileActionAriaLabel = hasRankingProfile && status.display_name
@@ -806,132 +794,14 @@ function LeaderboardCard({
           description={t(scope === 'local' ? 'ranking.local_empty_description' : 'ranking.empty_description')}
         />
       ) : (
-        <div className={styles.leaderboardResults}>
-          <div className={styles.podiumGrid} aria-label={`${t('ranking.rank')} 1–3`} data-ranking-podium>
-            {podium.map((entry, index) => (
-              <PodiumCard
-                key={entry.participant_id}
-                entry={entry}
-                position={index + 1}
-                metric={metric}
-                scope={scope}
-                onEditLocalProfile={onEditLocalProfile}
-                t={t}
-              />
-            ))}
-          </div>
-          {tableRows.length > 0 ? (
-            <div className={styles.tableScroll}>
-              <table className={styles.table}>
-                <thead>
-                  <tr>
-                    <th className={styles.rankColumn} data-ranking-rank-column>{t('ranking.rank')}</th>
-                    <th className={styles.participantColumn} data-ranking-participant-column>
-                      {t(scope === 'local' ? 'ranking.api_key' : 'ranking.participant')}
-                    </th>
-                    {metric === 'overall' ? (
-                      <>
-                        <th className={styles.numberCell}>{t('ranking.score')}</th>
-                        {OVERALL_METRICS.map((item) => <th key={item} className={styles.numberCell}>{t(`ranking.metric_short_${item}`)}</th>)}
-                      </>
-                    ) : <th className={styles.numberCell}>{t(`ranking.metric_${metric}`)}</th>}
-                  </tr>
-                </thead>
-                <tbody>
-                  {tableRows.map((entry, index) => (
-                    <tr key={entry.participant_id} data-ranking-row>
-                      <td className={styles.rankColumn} data-ranking-rank-column>
-                        <span className={styles.rankBadge} data-ranking-position>{index + 1}</span>
-                      </td>
-                      <td className={styles.participantColumn} data-ranking-participant-column>
-                        <div className={styles.participantCell}>
-                          <LeaderboardEntryAvatar
-                            entry={entry}
-                            scope={scope}
-                            className={styles.tableAvatar}
-                            onEditLocalProfile={onEditLocalProfile}
-                            t={t}
-                          />
-                          <strong>{entry.display_name}</strong>
-                        </div>
-                      </td>
-                      {metric === 'overall' ? (
-                        <>
-                          <td className={`${styles.numberCell} ${styles.scoreCell}`.trim()}>{formatLeaderboardValue(metric, entry, scope)}</td>
-                          {OVERALL_METRICS.map((item) => (
-                            <td key={item} className={styles.numberCell}>{formatOverallMetricValue(item, entry)}</td>
-                          ))}
-                        </>
-                      ) : <td className={`${styles.numberCell} ${styles.scoreCell}`.trim()}>{formatLeaderboardValue(metric, entry, scope)}</td>}
-                    </tr>
-                  ))}
-                </tbody>
-              </table>
-            </div>
-          ) : null}
-        </div>
+        <RankingLeaderboardResults
+          scope={scope}
+          metric={metric}
+          entries={rows}
+          onEditLocalProfile={onEditLocalProfile}
+        />
       )}
     </article>
-  );
-}
-
-function PodiumCard({ entry, position, metric, scope, onEditLocalProfile, t }: {
-  entry: RankingLeaderboardEntry;
-  position: number;
-  metric: RankingMetric;
-  scope: RankingScope;
-  onEditLocalProfile: (entry: RankingLeaderboardEntry) => void;
-  t: Translate;
-}) {
-  const value = formatLeaderboardValue(metric, entry, scope);
-  const valueSizeClass = value.length >= 11
-    ? styles.podiumValueCompact
-    : value.length >= 8
-      ? styles.podiumValueMedium
-      : '';
-
-  return (
-    <article
-      className={`${styles.podiumCard} ${styles[`podiumCard${position}` as keyof typeof styles]}`.trim()}
-      data-ranking-podium-rank={position}
-    >
-      <div className={styles.podiumRank}>
-        <span>{t('ranking.rank')}</span>
-        <strong>{String(position).padStart(2, '0')}</strong>
-      </div>
-      <LeaderboardEntryAvatar
-        entry={entry}
-        scope={scope}
-        className={styles.podiumAvatar}
-        onEditLocalProfile={onEditLocalProfile}
-        t={t}
-      />
-      <strong className={styles.podiumName}>{entry.display_name}</strong>
-      <span className={`${styles.podiumValue} ${valueSizeClass}`.trim()}>{value}</span>
-    </article>
-  );
-}
-
-function LeaderboardEntryAvatar({ entry, scope, className, onEditLocalProfile, t }: {
-  entry: RankingLeaderboardEntry;
-  scope: RankingScope;
-  className: string;
-  onEditLocalProfile: (entry: RankingLeaderboardEntry) => void;
-  t: Translate;
-}) {
-  if (scope !== 'local') {
-    return <RankingAvatar avatarID={entry.avatar_id} name={entry.display_name} className={className} decorative />;
-  }
-  return (
-    <button
-      type="button"
-      className={`${styles.localProfileAvatarButton} ${className}`.trim()}
-      aria-label={t('ranking.local_profile_edit_label', { name: entry.display_name })}
-      onClick={() => onEditLocalProfile(entry)}
-      data-ranking-local-profile-edit={entry.participant_id}
-    >
-      <RankingAvatar avatarID={entry.avatar_id} name={entry.display_name} decorative />
-    </button>
   );
 }
